ver. 1.1.2/20181221

+++ hois_back/src/main/java/ee/hitsa/ois/web/PersonController.java
@@ -16,6 +16,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import ee.hitsa.ois.domain.Person;
 import ee.hitsa.ois.domain.User;
+import ee.hitsa.ois.enums.Permission;
+import ee.hitsa.ois.enums.PermissionObject;
 import ee.hitsa.ois.exception.AssertionFailedException;
 import ee.hitsa.ois.service.PersonService;
 import ee.hitsa.ois.service.security.HoisUserDetails;
@@ -42,18 +44,21 @@ public class PersonController {
     @PostMapping
     public PersonWithUsersDto create(HoisUserDetails user, @Valid @RequestBody PersonForm request) {
         UserUtil.assertIsMainAdminOrSchoolAdmin(user);
+        UserUtil.assertHasPermission(user, Permission.OIGUS_M, PermissionObject.TEEMAOIGUS_KASUTAJA);
         return PersonWithUsersDto.of(personService.create(request), null);
     }
 
     @PutMapping("/{id:\\d+}")
     public PersonWithUsersDto save(HoisUserDetails user, @WithVersionedEntity(versionRequestBody = true) Person person, @Valid @RequestBody PersonForm personForm) {
         UserUtil.assertIsMainAdminOrSchoolAdmin(user);
+        UserUtil.assertHasPermission(user, Permission.OIGUS_M, PermissionObject.TEEMAOIGUS_KASUTAJA);
         return get(user, personService.save(personForm, person));
     }
 
     @GetMapping("/{id:\\d+}")
     public PersonWithUsersDto get(HoisUserDetails user, @WithEntity Person person) {
         UserUtil.assertIsMainAdminOrSchoolAdmin(user);
+        UserUtil.assertHasPermission(user, Permission.OIGUS_V, PermissionObject.TEEMAOIGUS_KASUTAJA);
         Set<User> users = person.getUsers();
         if (user.isSchoolAdmin()) {
             users = users.stream().filter(s -> s.getSchool() != null).filter(s -> EntityUtil.getId(s.getSchool()).equals(user.getSchoolId())).collect(Collectors.toSet());
@@ -64,6 +69,10 @@ public class PersonController {
     @GetMapping("/{person:\\d+}/users/{id:\\d+}")
     public UserDto getUser(HoisUserDetails userDetails, @WithEntity("person") Person person, @WithEntity User user) {
         UserUtil.assertIsMainAdminOrSchoolAdmin(userDetails);
+        UserUtil.assertHasPermission(userDetails, Permission.OIGUS_V, PermissionObject.TEEMAOIGUS_KASUTAJA);
+        if (!userDetails.isMainAdmin()) {
+            UserUtil.assertSameSchool(userDetails, user.getSchool());
+        }
         if (!EntityUtil.getId(person).equals(EntityUtil.getId(user.getPerson()))) {
             throw new AssertionFailedException("Person and user don't match");
         }
@@ -73,12 +82,14 @@ public class PersonController {
     @GetMapping("/{person:\\d+}/users")
     public UserDto initialValueForUser(HoisUserDetails user, @WithEntity("person") Person person) {
         UserUtil.assertIsMainAdminOrSchoolAdmin(user);
+        UserUtil.assertHasPermission(user, Permission.OIGUS_V, PermissionObject.TEEMAOIGUS_KASUTAJA);
         return personService.initialValueForUser(user, person);
     }
 
     @PostMapping("/{person:\\d+}/users")
     public UserDto createUser(HoisUserDetails user, @WithEntity("person") Person person, @Valid @RequestBody UserForm userForm) {
         UserUtil.assertIsMainAdminOrSchoolAdmin(user);
+        UserUtil.assertHasPermission(user, Permission.OIGUS_M, PermissionObject.TEEMAOIGUS_KASUTAJA);
         if (!user.isMainAdmin()) {
             userForm.setSchool(new EntityConnectionCommand(user.getSchoolId()));
         }
@@ -89,6 +100,7 @@ public class PersonController {
     @PutMapping("/{person:\\d+}/users/{id:\\d+}")
     public UserDto saveUser(HoisUserDetails userDetails, @WithEntity("person") Person person, @WithEntity User user, @Valid @RequestBody UserForm userForm) {
         UserUtil.assertIsMainAdminOrSchoolAdmin(userDetails);
+        UserUtil.assertHasPermission(userDetails, Permission.OIGUS_M, PermissionObject.TEEMAOIGUS_KASUTAJA);
         UserUtil.assertUserBelongsToPerson(user, person);
         UserUtil.assertCanUpdateUser(userForm.getRole());
         if (!userDetails.isMainAdmin()) {
@@ -101,14 +113,19 @@ public class PersonController {
     @DeleteMapping("/{id:\\d+}")
     public void deletePerson(HoisUserDetails user, @WithEntity Person person) {
         UserUtil.assertIsMainAdminOrSchoolAdmin(user);
+        UserUtil.assertHasPermission(user, Permission.OIGUS_M, PermissionObject.TEEMAOIGUS_KASUTAJA);
         personService.delete(user, person);
     }
 
     @DeleteMapping("/{person:\\d+}/users/{id:\\d+}")
     public void deleteUser(HoisUserDetails userDetails, @WithEntity("person") Person person, @WithEntity User user) {
         UserUtil.assertIsMainAdminOrSchoolAdmin(userDetails);
+        UserUtil.assertHasPermission(userDetails, Permission.OIGUS_M, PermissionObject.TEEMAOIGUS_KASUTAJA);
         UserUtil.assertUserBelongsToPerson(user, person);
         UserUtil.assertCanUpdateUser(EntityUtil.getCode(user.getRole()));
+        if (!userDetails.isMainAdmin()) {
+            UserUtil.assertSameSchool(userDetails, user.getSchool());
+        }
         personService.deleteUser(userDetails, user);
     }
 }
